requirements.txt
# Core ML/DL Frameworks
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0

# MLOps Tools
mlflow>=2.8.0
optuna>=3.4.0

# API and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# CLI and Utils
typer>=0.9.0
rich>=13.6.0
tqdm>=4.66.0

# Testing and Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0

# Monitoring
prometheus-client>=0.18.0
psutil>=5.9.0

# Data Versioning (optional)
dvc>=3.30.0

# Additional
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
